ai-service/requirements.txt

fastapi==0.110.0
uvicorn==0.27.1
google-generativeai==0.3.2
python-multipart==0.0.9
websockets==12.0
numpy==1.26.4
pyaudio==0.2.14
soundfile==0.12.1
requests==2.31.0
pydub==0.25.1
httpx
google-cloud-texttospeech
